refactor(link-daemon): rename sandbox-provider.ts → user-desktop-provider.ts
The file implements the desktop-side provider for the user-desktop kind;
the new name is self-describing within the link-daemon directory.
