+++ apps/mesh/src/link-daemon/user-desktop-provider.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from "bun:test";
 import { mkdtempSync, rmSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import { createDesktopSandboxProvider } from "./sandbox-provider";
+import { createDesktopSandboxProvider } from "./user-desktop-provider";
 
 function fakeDaemonSpawner() {
   return {
